requirements.txt
flask
flask-cors
groq
python-dotenv
requests
